backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
requests==2.31.0
python-ollama==0.2.4
psutil==5.9.5
pytest==7.3.1
pytest-cov==4.1.0
python-dateutil==2.8.2 